Fin de la BDD

+++ PokePote/app/src/main/java/com/example/martin/pokepote/PKM_BDD.java
@@ -1,14 +1,11 @@
 package com.example.martin.pokepote;
 
-/**
- * Created by Shiita on 23/05/2015.
- */
 public class PKM_BDD {
-    long ID;
+    Integer ID;
     Integer favori;
     Integer equipe;
 
-    public PKM_BDD(long ID, Integer favori, Integer equipe) {
+    public PKM_BDD(Integer ID, Integer favori, Integer equipe) {
         this.ID = ID;
         this.favori = favori;
         this.equipe = equipe;
